Self-reported cme form expects creditType param.
For sr-cme, the value of creditType must be either:
'1': for Category 1 Credit
'0': for Other

The entries in the feed have an extra field called creditType.
Value is always '1' for browser-cme.
Value is either '1' or '1' for sr-cme.
Value is empty '' for other entry types.

The login response has a new key called creditTypes
which returns a list of dicts for the possible options
for creditType for sr-cme.