Remove unused cmeTags from user on profile update.
If user selects a practice specialty, the tags of the
pracspec are assigned to the user, and are then available
to the user for tagging their entries.
But if the user removes a pracspec, then for each tag
assigned to the pracspec: check if the tag is unused (not
associated with any of the user's entries). If tag is
unused, then delete the user-tag association.

settings.py: remove logdna handler from the gen/api/mgmt loggers.