Schema change:suggested cmeTags on BrowserCmeOffer
New M2M field on BrowserCmeOffer to store suggested tags.
This must be further processed by the UI (by intersecting
with the user's allowed cmeTags) to pre-select certain
tags in the form to redeem the offer.

New field on EligibleSite: all_specialties booleanField.
Flag to state that the site is associated with all pracspecs
by default.