profile.cmeTags is list of objects in serializer.
1. When profile is updated, the cmeTags of the serialized
updated instance is now returned as a list of objects (per
request of GlebS).

2. In auth_status and login response, the profile.cmeTags
is also a list of objects. This means we no longer need the
extra key cmeTags because it contains redundant data. The
cmeTags key was removed from auth_status and login response.
They use the ReadProfileSerializer.

3. ProfileDetail endpoint renamed to ProfileUpdate endpoint.
It now only is used to update a profile instance. It uses
the UpdateProfileSerializer.

4. The UpdateProfileSerializer checks if the specialties key
is present in the validated_data before doing any calculations.
This was done to handle the PATCH case in which not all keys
are present in the validated_data.