Merge invitation promo feature (#7)
* Updated .gitignore

* Add endpoint invalidate-entry/pk/

This is used by the UI to "delete" an entry.

* Set min_length for profile.inviteId to 5 chars.

Extend Hashids alphabet with some extra chars to enable more
compaction.
auth_views: If inviteid is passed a GET parameter to login_via_token
then it is passed along to authenicate as inviterId.
auth_backend: Set inviter on profile if valid inviterId is given.
Also specify alphabet for Hashids and set min_length=5.

* sumBrowserCme filters by entry__valid=True.

The filter kwarg should be entry__valid (not plain valid).

* Add BooleanField valid to BrowserCmeOffer.

default value is True. If user invalidates a br-cme entry, then the
action will also invalidate the offer.

* Fixed typo in attribute name

* Schema change: Add models for invitation discount.

New models: Discount and InvitationDiscount.
Discount = list of available discounts
InvitationDiscount: stores inviter-invitee associations.

payment_views.NewSubscription: changed name of the POST parameters to
use underscore instead of hyphen (e.g. do-trial is now do_trial).
The view is now a CreateAPIView and uses a serializer_class.

auth_views: The login context now uses ReadUserSubsSerializer
for the subscription object. It has an extra key called invitation.
By default, the value is None. It finds the latest (last modified)
InvitationDiscount for the user as inviter. If one exists, then:
1. The object is stored in the sub-key called latestInvitation and the
value is from ReadInvitationDiscountSerializer.
2. The sub-key totalCredit contains the total amount of credit earned
toward their next full-price subscription (credits do not apply to first
year because that already has the plan first year discount).

* auth_views: add totalCompleteInvites to invitation

Remove latestInvitation object and add totalCompleteInvites
integer instead. It equals the total number of completed
InvitationDiscount instances for which the user is inviter
(each of these invitees has begin an Active subscription).

* models.py: Bug fix in createBtSubscription.

Also, check if need to apply invitee_discount in
SwitchTrialToActive in the case of: invitee signs up in Trial but then
switches to Active. The original Trial subs is canceled, so the
invitee_discount can be applied to the new Active subs.

* Public endpoint to get firstName from inviteId

Example: GET api/v1/invite-lookup/abcde/ will attempt to look up
the user for inviteId=abcde. If a user is found, it returns an
object {username: string}.  If none found, it returns 404.

* models.py fix invitee_discount in SwitchTrialToActive.

* Add bash scripts to be run by cron task

Scrips to run inviterDiscount management command.

* Bash script to run handlePaymentTransactions cron.