SubscriptionTransaction model: add trans_type field.
A BT transaction has type sale or credit.  Add trans_type CharField to
SubscriptionTransaction model to store this info. If type is credit,
this means the amount was credited back to the user. In this case,
the BT processor_authorization_code is None which is converted to empty
str for proc_auth_code value in db.