Add is_active to cmeTags in Profile serializers.
In the ReadProfileSerializer and UpdateProfileSerializer,
the cmeTags is now a list of objects where each tag object
has been annotated with an is_active:bool attribute.
If the tag has is_active=True this means the user is currently
using it (via their current practice specialties), else it
is False.