Schema change: add Affiliate-related models.
New public endpoint aff-lookup/<affiliateId> which returns
the username (from Profile), and personalText/photoUrl from Affiliate.

Affiliate: one-to-one field with User model. An affiliate must be a
user in the db and have a profile instance.

AffiliatePayout stores the affiliate-convertee relations (the users
that the affiliate has introduced to OrbitCme). A payout is added
once the new user has begun an Active subs.

BatchPayout: to be used along with paypal payout API to deliver
the payouts to the affiliates.