Add endpoint to allow user to manage cmeTags.
Schema change: profile.cmeTags is now stored in a new model: ProfileCmetag
which is the many-to-many through relation. The is_active field is now an
actual column in this table.

New endpoint /profiles/set-cmetags/ is to allow user to control the
is_active flag on the cmeTags assigned to their profile. This endpoint
does not create or delete any tag associations, it only modifies the
is_active flag of existing tags.

ProfileSerializer (Read/Update): the 'id' key in the cmeTags list is
renamed to 'tag_id' because the value is the pkeyid of the tag.