Monthly earnings statement email to each affiliate.
For each Affiliate who earned a payout, the updateAfflPayout mgmt cmd
calls sendAfflEarningsStatementEmail to send an email containing a
summary of their total conversions and commission for the past month.
This email is in addition to the actual payout email sent by PayPal.

For each Affiliate who did not earn a payout, the afflPayout mgmt
command calls sendAfflConsolationEmail to send an email stating that
they did not receive a payout this month.