New models and endpoints for Orbit Stories.
New required entryType row must be added to DB for: story-cme
New models: Story, and StoryCme which is a subclass of Entry model.

To create a new Story, use the Admin interface.
For testing, use Swagger to create a StoryCme entry.

Endpoints:
StoryDetail: returns info about the latest non-expired Story with a
launch_url customized for the user.