New endpoints for upgrade plan
UpgradePlanAmount: GET method that returns the amount the user will
be charged for upgrading to the higher-priced plan. It expects the
plan pk as a URL parameter.
UpgradePlan: POST method that accepts plan pk and either a nonce or
token. It will cancel the user's current subscription and begin a new
subscription on the new plan.