Changes to UpgradePlan/UpgradePlanAmount endpoints.
Changes to UpgradePlan and UpgradePlanAmount to handle cases
of when current user subscription is:
Expired: user owes full price on the new plan
Trial/Trial-Canceled: user is eligible for any signup discounts
Active-Canceled: user is eligible for proration on first-year discount
but does not get any earned discounts (that is for Active users only).
Also, UpgradePlan only cancels the current subscription if it is not
already in a terminal state.