Add management cmd to complete scheduled downgrades.
When user requests to downgrade, the status is changed to
UI_ACTIVE_DOWNGRADE, and the user remains in the upgraded plan until the
end of their billing cycle. This is done because they already paid for
the cycle and we do not offer refunds.
New management command: completeDowngradeSubs terminates the old
subscription when it is close to the billingEndDate and starts new
subscription on the downgraded plan using existing payment token for the
customer. This completes the scheduled downgrade.
Schema change: Add SubscriptionEmail model to track the sending of
various emails (such as card expired). Also, add next_plan FK on
UserSubscription model.