Admin interface: use django-autocomplete-light (dal)
Install dal package and use it in the admin interface for
improving the dropdown menu to select a user in the AuthImpersonation
form.
ac_views.py: The widget requires an endpoint that filters the queryset
based on the search term.
Note: need to shift-refresh browser to force to update its js cache for
the admin site.