login endpoint expect plan GET parameter on signup
auth_views: For new user signup, the login endpoint expects a GET
parameter called plan which contains the planId of the selected plan.

auth_backends: In new user creation, the planId is stored in the
profile, and degree/specialty initial value is set from the plan-key.

payment_views: NewSubscription endpoint no longer expect plan in the
input data. Instead, it uses profile.planId to get the plan.

admin: Auto-set planId based on plan name and hashid of next pk. This is
done by overriding the save method of the ModelForm used by the Admin.