Schema change: add nullable entry to UserFeedback
UserFeedback model: add nullable OneToOneField to Entry model. If one is
provided, the feedback is associated with a particular entry.

BrowserCme model: Add CharField for planText. This is an optional field
to store the explanation for change in clinical plan provided by the
user when redeeming the offer.

serializers:
UserFeedback: it expects a field called entry in the POST data. The
value should be either null or a valid entry pk.
BrowserCme: added planText to the read and create serializers.