afflPayout mgmt cmd: fix sending duplicate email
Consolation emails to associates who did not earn any payout in the
month are sent with report_only option, since they can be sent at that
time. When afflPayout is called to submit an actual batch payout, the
consolation emails will not be re-sent.
emailutils: Handle blank/stale affiliateId stored in convertee profile.