Change output serializer for CreateBrowserCme endpoint.
Replace credits in the response dict with key: brcme. The value is the
brcme model instance converted to dict. This allows the UI to have the
full brcme object with any default values populated by the server.

renewal_reminder.html: bug fix in nextBillingAmount var. Also, remove
link to feedback.