inviterDiscount mgmt cmd: skip inviter on free plan.
inviterDiscount: Skip processing InvitationDiscount if inviter is not on
a paid plan.

payment_views: profile.affiliateId must be set in order for convertee
discount to be True. It is not enough to check that profile.inviter is
an Affiliate, the profile.affiliateId must also be set. If it is not
set, this means the inviter used his inviteURL rather than his
affiliateId to convert the user.

models/subscription: If bt_subs is successfully created, we must return
without raising any uncaught Exception or else the transaction will
rollback and the user_subs will not be committed to the db. Put the
InvitationDiscount/AffiliatePayout creation code inside a try/catch
block to handle any uncaught exception. It is essential that user_subs
remains in sync with the bt_subs even if ID/AFP is not created.