WIP Draft of email sync architecture
This commit introduces functionality to sync a contact list between a local database and a list in an Email Service Provider, such as Mailchimp. At this time only Mailchimp is supported, but other ESPs could be supported by inheriting from the base class EspApiBackend and then adding code specific to that ESP API.

At this point in time, only a handful of fields (name, email, user_id) get synched, but more could easily be added.