Rename main file to suit conventions, notify if mandatory field data is missing
This commit changes the name of the main file to suit local conventions, from camelCase to snake_case.
It also checks to see if local users to update have all fields mandatory by ESP, and notifies if any users are lacking required data.
Finally, the requirement of having the necessary variables ORBIT_MAILCHIMP_USERNAME and ORBIT_MAILCHIMP_API_KEY in the .env file is added to the Readme, and an option to add a default Email Service Provider is moved to settings.