Organization model: add BooleanField activateGoals
Default value is True. If True: members of this enterprise
org will have usergoal/compliance checking enabled. If False,
they will still see the Track view in the UI, but it will be empty.
Note: a manual command must be run to either create or delete usergoals
if org.activateGoals is changed at a later date.