New endpoint: /goals/create-license
UserGoalSummary: add licenseNumber to response for user license goals.
CreateUserLicenseGoal endpoint is for enterprise admin to add a new
StateLicense for a user in their org. It creates a new StateLicense, new
LicenseGoal and any applicable credit goals.
UpdateUserLicenseGoal: changed permissions on endpoint to allow it to be
called by an enterprise admin, or the user who owns the license.