New endpoint: /enterprise/report
This endpoint lists the available reports from the OrgReport model.
Schema changes:
OrgReport: table to store list of available reports for an orgadmin to
download. For each report, server sends the endpoint url that should be
called. The endpoint url will return JSON and the UI will generate a csv
file.
OrgFile: add file_type field to categorize the file.