New endpoint: /enterprise/upload-license to upload license file
This endpoint accepts uploaded file for updating either DEA or State
Licenses for existing users (it does not handle creating new users).
It determines the file_type and pre-validates the file. It returns
some validation results as the response. It does not actually process
the file (which must be done after the end user clicks confirm).