Add management command for makeTuftsIntMedReport.
This mgmt cmd should be called daily by a crontask. It will only
activate twice a month and send an email to Tufts with the
report data for the Internal Medicine providers who have completed
their profile. If no data exists for the time period, it still sends
a notification email to Tufts.