auth_views: trying to fix new user signup
The function jwt_get_username_from_payload will return an invalid
payload error to the client if authenticate  does not return a
user. But it is unable to pass the request object to authenticate
because it doesnt receive it as an arg!
So Auth0Backend.authenticate method create a new User with username=auth0Id
and returns the user. This takes care of jwt_get_username_from_payload
returning a valid user.
auth_views: signup function gets the access_token from the request header,
and passes it to auth0 python api to get the user info for the token.
Then it calls auth_backends.configure_user with the user info to complete
the signup.
auth_backends: configure_user updates the User instance that was created
by authenticate and sets email/username to the corect email. It also
creates the Profile and completes the initialization.
urls.py: comment out swagger because it is not compatible with new way of
authentication. Will re-visit this later.