admin.py: Add help_text for the Plantags section in the PlanForm.
Note: when using TabularInline formset, Django does not display
the help_text for the form fields (it is only available as a tooltip
where you have to hover over the question mark with your mouse).

So, I added my own help text to the verbose_name_plural field of
the PlantagInline class. This text is automatically wrapped inside
an h2 tag by Django (and auto-uppercased it seems).