enhanced weekly email version 1
Ram's enhanced version of naoki's weekly email version 1 - adds html version (no plaintext version), including header. Based on example titled "Here’s an example of how to create an HTML message with an alternative plain text version. To make things a bit more interesting, we include a related image in the html part, and we save a copy of what we are going to send to disk, as well as sending it." from: https://docs.python.org/3/library/email.examples.html